Initial commit to get releases working
This should cause tagged commits to trigger the github
releases integration in travis, with each release having
a single binary and a .sha256 file whose contents are the
checksum of the binary.

I added release and release-all scripts from the nebula-sdk
directory and began reworking them for this tool; they are
not used (yet?) because there's about 2 or 3 more layers of
abstraction than we need.